requirements.txt
shiny
shinywidgets
plotly
pandas
scipy
supabase
python-dotenv
openpyxl
kaleido